requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
celery>=5.3.6
redis>=5.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
requests>=2.31.0
pdfminer.six>=20231228
python-dotenv>=1.0.1
httpx>=0.26.0
